sorting algorith

+++ sort.h
@@ -29,8 +29,8 @@ void swap(listint_t *node, listint_t *next, listint_t **list);
 void selection_sort(int *array, size_t size);
 
 void quick_sort(int *array, size_t size);
-void q_sort(int *array, size_t size, size_t low, size_t high);
-size_t partition(int *array, size_t size, size_t low, size_t high);
+void q_sort(int *array, size_t size, int low, int high);
+int partition(int *array, size_t size, int low, int high);
 void swap_index(int i, int j);
